+++ sdk/keyvault/azure-mgmt-keyvault/azure/mgmt/keyvault/aio/_key_vault_management_client.py
@@ -75,7 +75,7 @@ class KeyVaultManagementClient:  # pylint: disable=too-many-instance-attributes
     :type subscription_id: str
     :param base_url: Service URL. Default value is None.
     :type base_url: str
-    :keyword api_version: Api Version. Default value is "2024-11-01". Note that overriding this
+    :keyword api_version: Api Version. Default value is "2025-05-01". Note that overriding this
      default value may result in unsupported behavior.
     :paramtype api_version: str
     :keyword int polling_interval: Default waiting time between two polls for LRO operations if no