+++ sdk/keyvault/azure-mgmt-keyvault/generated_samples/get_key_version.py
@@ -40,6 +40,6 @@ def main():
     print(response)
 
 
-# x-ms-original-file: specification/keyvault/resource-manager/Microsoft.KeyVault/stable/2024-11-01/examples/getKeyVersion.json
+# x-ms-original-file: specification/keyvault/resource-manager/Microsoft.KeyVault/stable/2025-05-01/examples/getKeyVersion.json
 if __name__ == "__main__":
     main()