+++ sdk/keyvault/azure-mgmt-keyvault/generated_samples/list_vault.py
@@ -36,6 +36,6 @@ def main():
         print(item)
 
 
-# x-ms-original-file: specification/keyvault/resource-manager/Microsoft.KeyVault/stable/2024-11-01/examples/listVault.json
+# x-ms-original-file: specification/keyvault/resource-manager/Microsoft.KeyVault/stable/2025-05-01/examples/listVault.json
 if __name__ == "__main__":
     main()