+++ sdk/keyvault/azure-mgmt-keyvault/generated_tests/test_key_vault_management_mhsm_private_endpoint_connections_operations_async.py
@@ -25,7 +25,7 @@ class TestKeyVaultManagementMHSMPrivateEndpointConnectionsOperationsAsync(AzureM
         response = self.client.mhsm_private_endpoint_connections.list_by_resource(
             resource_group_name=resource_group.name,
             name="str",
-            api_version="2024-11-01",
+            api_version="2025-05-01",
         )
         result = [r async for r in response]
         # please add some check logic here by yourself
@@ -38,7 +38,7 @@ class TestKeyVaultManagementMHSMPrivateEndpointConnectionsOperationsAsync(AzureM
             resource_group_name=resource_group.name,
             name="str",
             private_endpoint_connection_name="str",
-            api_version="2024-11-01",
+            api_version="2025-05-01",
         )
 
         # please add some check logic here by yourself
@@ -77,7 +77,7 @@ class TestKeyVaultManagementMHSMPrivateEndpointConnectionsOperationsAsync(AzureM
                 "tags": {"str": "str"},
                 "type": "str",
             },
-            api_version="2024-11-01",
+            api_version="2025-05-01",
         )
 
         # please add some check logic here by yourself
@@ -91,7 +91,7 @@ class TestKeyVaultManagementMHSMPrivateEndpointConnectionsOperationsAsync(AzureM
                 resource_group_name=resource_group.name,
                 name="str",
                 private_endpoint_connection_name="str",
-                api_version="2024-11-01",
+                api_version="2025-05-01",
             )
         ).result()  # call '.result()' to poll until service return final result