+++ sdk/keyvault/azure-mgmt-keyvault/generated_tests/test_key_vault_management_secrets_operations.py
@@ -41,7 +41,7 @@ class TestKeyVaultManagementSecretsOperations(AzureMgmtRecordedTestCase):
                 },
                 "tags": {"str": "str"},
             },
-            api_version="2024-11-01",
+            api_version="2025-05-01",
         )
 
         # please add some check logic here by yourself
@@ -68,7 +68,7 @@ class TestKeyVaultManagementSecretsOperations(AzureMgmtRecordedTestCase):
                 },
                 "tags": {"str": "str"},
             },
-            api_version="2024-11-01",
+            api_version="2025-05-01",
         )
 
         # please add some check logic here by yourself
@@ -81,7 +81,7 @@ class TestKeyVaultManagementSecretsOperations(AzureMgmtRecordedTestCase):
             resource_group_name=resource_group.name,
             vault_name="str",
             secret_name="str",
-            api_version="2024-11-01",
+            api_version="2025-05-01",
         )
 
         # please add some check logic here by yourself
@@ -93,7 +93,7 @@ class TestKeyVaultManagementSecretsOperations(AzureMgmtRecordedTestCase):
         response = self.client.secrets.list(
             resource_group_name=resource_group.name,
             vault_name="str",
-            api_version="2024-11-01",
+            api_version="2025-05-01",
         )
         result = [r for r in response]
         # please add some check logic here by yourself